Fix(?) contradiction(?) in model loading
Vertex counts are loaded at uint32 in the model header, but the vertex
index is read as int16 when parsing the model. This seems like a
contradiction, as it would allow the file to contain 4M vertices but
only allow some 37k of them to be indexable. There are some parsing
errors with this related to an invalid index, so it seems plausible
that this could be the cause.