Separate menu entries and parsing functions for SH3 MDL files
SH3 MDL files have a different structure to SH2 MDL
files and so cannot be parsed in the same way. Separate menu entries
are therefore required, and different parsing functions are needed.

So far, a modified `parse_morph_targets` has been created for SH3
models (with understanding of the difference in the formats gleaned
from https://github.com/alanm20/SH3_chr), but parsing of the submeshes
and the textures is not yet possible. These aspects of parsing have
been disabled until they can be fixed for SH3 meshes.